refactor: ProjectDetail 폴더 Project 폴더 안으로 이동
